Rename feedback-page to feedback
None of the other pages have a 'page' suffix
