+++ app/full-page-examples.js
@@ -1,5 +1,5 @@
 module.exports = (app) => {
-  require('./views/full-page-examples/feedback-page')(app)
+  require('./views/full-page-examples/feedback')(app)
   require('./views/full-page-examples/how-do-you-want-to-sign-in')(app)
   require('./views/full-page-examples/news-and-communications')(app)
   require('./views/full-page-examples/passport-details')(app)
